#197 💎 merge getMostSpecificPathFromError + getValidationMessageFromPath
