+++ src/utils/validate/index.ts
@@ -1,8 +1,7 @@
 export * from './baseUrlSchema/baseUrlSchema';
 export * from './corsSchema/corsSchema';
 export * from './databaseConfigSchema/databaseConfigSchema';
-export * from './getMostSpecificPathFromError';
-export * from './getValidationMessageFromPath';
+export * from './getValidationMessage';
 export * from './graphqlConfigSchema/graphqlConfigSchema';
 export * from './interceptorsSchema/interceptorsSchema';
 export * from './portSchema/portSchema';
