+++ src/utils/validate/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './checkModeSchema/checkModeSchema';
 export * from './entitiesSchema/entitiesSchema';
+export * from './oneKeyDispatchSchema/oneKeyDispatchSchema';
 export * from './sharedSchema/sharedSchema';
